requirements.txt
FastMCP
azure-kusto-data